Changes in this commit: 1 - Finishing movingToDelivered() function. 2 - Changing the assignedCargos dataStructure to PriorityQueue. 3 - Changing in assigningCargos() functions.

+++ PriorityNode.h
@@ -6,7 +6,7 @@ class PriorityNode
 	T item;
 	PriorityNode<T>* next;
 public:
-	int priority;
+	float priority;
 	void setItem(const T& r_Item);
 	void setNext(PriorityNode<T>* nextNodePtr);
 	T getItem() const;
